@@ -5,12 +5,12 @@
   "private": true,
   "dependencies": {
     "moment": "^2.25.3",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
     "react-markdown": "^4.3.1",
-    "react-router-dom": "^5.1.2",
-    "react-scripts": "^4.0.0",
+    "react-router-dom": "^6.2.1",
+    "react-scripts": "^4.0.3",
     "react-share": "^4.4.0",
     "react-syntax-highlighter": "^15.4.5",
     "remark-gfm": "^3.0.1",
