@@ -2,8 +2,11 @@ import styled from "styled-components";
 
 export const BlogWrapper = styled.div`
   text-align: left;
+  color: var(--blogColor);
+  padding-top: 20px;
+  padding-bottom: 20px;
   h2 {
-    color: rgb(105, 102, 102);
+    color: var(--blogHeaderColor);
   }
   li:nth-child(even) {
     color: #94095f;
@@ -19,9 +22,36 @@ export const BlogWrapper = styled.div`
 export const ShareButtonsContainer = styled.div`
   display: flex;
   justify-content: end;
-  align-items: center;
-  button {
+  text-align: center;
+
+  & > div {
     margin-left: 10px;
+  }
+  .count {
+    margin-top: -10px;
+  }
+  button {
     cursor: pointer;
   }
+  .react-share__ShareCount {
+    font-size: 9px;
+    color: gray;
+  }
+`;
+
+export const Button = styled.button`
+  border: 1px solid #145e98;
+  border: none;
+  padding: 3px 10px;
+  border-radius: 10px;
+  box-shadow: -3px -1px 5px 2px lightgray;
+  background: transparent;
+  color: #145e98;
+  margin-bottom: 10px;
+`;
+
+export const Aligner = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
 `;
