@@ -1,4 +1,3 @@
-import React from "react";
 import { Link } from "react-router-dom";
 import { blogs } from "../blogs/headline";
 import PageHeader from "../utils/Helmet";
@@ -7,8 +6,9 @@ import { BlogHeaders, BlogTitles } from "./styled";
 export default function Blogs() {
   return (
     <BlogHeaders>
-      <PageHeader title="All Blogs" />
-      <h3>All blogs</h3>
+      <PageHeader title="All Blogs - Ulan Rakymzhanov" />
+      <h2>All blogs</h2>
+      <hr />
       <BlogTitles>
         {blogs.map((blog) => {
           return (
@@ -18,9 +18,7 @@ export default function Blogs() {
                 pathname: `/blogs/${blog.link}`,
               }}
             >
-              <h4>
-                {blog.id}. {blog.title}
-              </h4>
+              <h3> ~ {blog.title}</h3>
             </Link>
           );
         })}
