@@ -1,16 +1,4 @@
-import styled from 'styled-components';
-
-// Router
-// export const MainRouter = styled.div`
-//   border: 1px solid;
-//   font-size: var(--mainFont);
-//   overflow: auto;
-//   min-height: 100vh;
-//   @media (max-width: 700px) {
-//     margin-left: 0;
-//     min-height: 100%;
-//   }
-// `;
+import styled from "styled-components";
 
 // BlogsNavigatorPage
 export const BlogHeaders = styled.div`
@@ -21,6 +9,9 @@ export const BlogHeaders = styled.div`
 export const BlogTitles = styled.div`
   text-align: left;
   a {
-    color: var(--darkred);
+    color: #000;
+  }
+  a:hover {
+    text-decoration: underline;
   }
 `;
