@@ -1,5 +1,5 @@
-import React from 'react';
-import styled from 'styled-components';
+import React from "react";
+import styled from "styled-components";
 
 const LayoutWrapper = (props) => {
   return <Div>{props.children}</Div>;
@@ -10,6 +10,6 @@ export default LayoutWrapper;
 const Div = styled.div`
   padding-bottom: 50px;
   padding-top: 60px;
-  padding-right: 20px;
-  padding-left: 20px;
+  padding-right: 25px;
+  padding-left: 25px;
 `;
